[NO-TICKET] Remove unused profiler PID controller
**What does this PR do?**

This PR removes the unused profiler PID controller implementation.

**Motivation:**

In #3190 we imported the PID controller, and planned to use it to
control the allocation profiler. In the end, we ended up going with
a different solution for that (#3395) so let's remove the PID controller
for now.

We can always revert this commit if we need it again.

**Additional Notes:**

N/A

**How to test the change?**

Validate that CI is still green :)