[PROF-8944] Upgrade to libdatadog 6
**What does this PR do?**

This PR upgrades dd-trace-rb to use libdatadog 6. The highlight of
this release are memory footprint improvements when using the
profiling timeline feature.

There was a small breaking API change -- the addition of the
`info` parameter in `ddog_prof_Exporter_Request_build`, which will
be used in #3357 but for this PR I only passed in a `NULL`.

**Motivation:**

Pick up the timeline memory improvements.

**Additional Notes:**

N/A

**How to test the change?**

Our existing test coverage includes libdatadog testing, so a green
CI is good here :)