[NO-TICKET] Add Ruby profiler benchmark with GC profiling
**What does this PR do?**

This PR adds a new configuration to for testing in the benchmarking
platform.

This configuration enables the "GC profiling" (on top of timeline),
and will be used to validate when we're good to enable this feature
by default.

**Motivation:**

Validate performance impact of "GC profiling" feature, especially when
used together with timeline.

**Additional Notes:**

I'm preparing a similar PR to add this configuration also to the
reliability environment configurations.

**How to test the change?**

Validate that benchmarking platform profiling runs now include a
variant where GC information shows up in the flamegraph, as well
as in the timeline visualization.