[PROF-8942] Remove separate benchmarking configurations for profiling timeline
**What does this PR do?**

As soon as we merge https://github.com/datadog/dd-trace-rb/pull/3428 the
profiler will default to timeline being enabled.

This means that the "profiler" configurations will all have timeline,
so having separate configurations for timeline does not make any sense.

**Motivation:**

Remove timeline testing configurations that no longer make sense.

**Additional Notes:**

For the configuration with GC profiling, I've removed the
`DD_PROFILING_EXPERIMENTAL_TIMELINE_ENABLED` environment variable,
since, again, timeline is on by default.

We could potentially keep a configuration without timeline around,
but we already have a lot of configurations, so for now I've chosen
not to do so.

**How to test the change?**

This can be tested by manually triggering a deploy to the
benchmarking platform.