[PROF-9147] Add workaround for Ruby VM bug causing crash in gc_finalize_deferred
**What does this PR do?**

This PR adds a workaround for a VM bug
(https://bugs.ruby-lang.org/issues/19991) that has affected at least
two customers (the other being
https://github.com/DataDog/dd-trace-rb/issues/3430)

The workaround is... not very pretty... but I believe will get the
job done, for the reasons explained in the big comment I added.

**Motivation:**

The good thing about this crash is that it seems extremely rare.

The bad thing is that without this workaround, only moving to Ruby 3.3+
is the fix, and we still have a lot of customers on older Ruby
versions.

We've documented in
https://docs.datadoghq.com/profiler/profiler_troubleshooting/ruby/#segmentation-faults-in-gc_finalize_deferred-in-ruby-versions-26-to-32
that the no signals workaround can be used to work around the issue
but:

a. Enabling it for everyone on < 3.3 is really crappy, as it yields
biased samples and the crash only happens in the presence of
specific types of Ruby objects (e.g. objects with finalizers) that
not all apps will run into.

b. For folks that are affected by this issue, it may take quite
a while for them to connect the bug to the profiler, since their
crash log will point at `gc_finalize_deferred` and won't mention
datadog at all.

**Additional Notes:**

N/A

**How to test the change?**

This is the million dollar question. I am unable to reproduce the
issue locally just by running the profiler.

I was able to simulate it by changing Ruby's `vm_trace.c` as follows:

```
@@ -1748,6 +1748,8 @@ rb_workqueue_register(unsigned flags, rb_postponed_job_func_t func, void *data)
     return TRUE;
 }

+static void dummy_method(void *_unused) { }
+
 void
 rb_postponed_job_flush(rb_vm_t *vm)
 {
@@ -1775,7 +1777,11 @@ rb_postponed_job_flush(rb_vm_t *vm)
             while ((index = vm->postponed_job_index) > 0) {
                 if (ATOMIC_CAS(vm->postponed_job_index, index, index-1) == index) {
                     rb_postponed_job_t *pjob = &vm->postponed_job_buffer[index-1];
-                    (*pjob->func)(pjob->data);
+                    rb_postponed_job_func_t the_func = (*pjob->func);
+                    if (the_func != dummy_method) rb_postponed_job_register_one(0, dummy_method, NULL);
+                    //if (the_func != dummy_method) rb_postponed_job_register_one(0, dummy_method, GET_VM()->objspace);
+                    (the_func)(pjob->data);
                 }
```

The idea being that it's equivalent to the profiler SIGPROF
interruption landing right between reading `pjob->func` and
`pjob->data`.

I then ran the following Ruby snippet (which triggers
`gc_finalize_deferred`):

```ruby
class UsesFinalizer
  DUMMY_FINALIZER = proc { }
  def initialize = ObjectSpace.define_finalizer(self, DUMMY_FINALIZER)
end
UsesFinalizer.new while true
```

...and this gets me a crash that matches the ones linked in
* https://bugs.ruby-lang.org/issues/19991
* https://github.com/DataDog/dd-trace-rb/issues/3430

If instead of passing `NULL`, we uncommend the second option that
passes in `GET_VM()->objspace`, the crash goes away.

Thus I believe that my fix works and is equivalent to what
the experiment above does.