[PROF-9257] Upgrade libdatadog to fix incorrect platform detection for x86_64-linux-gnu/aarch64-linux-gnu
**What does this PR do?**

This PR pulls in the change from
https://github.com/datadog/libdatadog/pull/338 that fixes
platform detection not working correctly when the user's platform is
`x86_64-linux-gnu`/`aarch64-linux-gnu`.

This manifested to a user as the following error:

> Profiling was requested but is not supported, profiling disabled: Your
> ddtrace installation is missing support for the Continuous Profiler
> because the `libdatadog` gem installed on your system is missing
> binaries for your platform variant.
> (Your platform: `x86_64-linux-gnu`)
> (Available binaries: `x86_64-linux`, `x86_64-linux-musl`)

The fix is entirely on the libdatadog side; on the dd-trace-rb side
we only need to pick up that libdatadog version.

**Motivation:**

Make sure our platform detection code is correct.

**Additional Notes:**

I've also slightly tweaked the error message to output the platform
as seen/normalized by libdatadog, as otherwise it could get
confusing to debug.

I'm opening this PR before the libdatadog PR is merged and
released, so CI will obviously be red (since the release is not
on rubygems.org). I'll re-trigger CI once the libdatadog release
is out. (I've tested it locally and it works fine)

**How to test the change?**

This change includes test coverage on the libdatadog side.