Replace `profiling.advanced.experimental_allocation_enabled` with `profiling.allocation_enabled`
As we are no longer considering this option experimental, let's
rename it to reflect that.

Furthermore, enabling allocation profiling is not considered an
"advanced" option.

In the process, let's deprecate the old
`profiling.advanced.experimental_allocation_enabled` and leave a note
asking customers to update their `Datadog.configuration` block.

I did spend some time trying to get the value from
`experimental_allocation_enabled` into the new setting automatically,
but because they are in separate options groups, I wasn't able to
find a way to access the new setting from the old one.

I was able to still support the old environment variable, so at least
we have that.