[NO-TICKET] Fix nightly test build deploys
**What does this PR do?**

This PR fixes the following error in our nightly deploys of test builds:

```
Unable to create pipeline
'install-dependencies-amd64: [2.7.8]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'install-dependencies-amd64: [3.0.6]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'install-dependencies-amd64: [3.1.4]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'install-dependencies-amd64: [3.2.2]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'install-dependencies-arm64: [2.7.8]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'install-dependencies-arm64: [3.0.6]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'install-dependencies-arm64: [3.1.4]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'install-dependencies-arm64: [3.2.2]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'package-amd64: [deb-x64]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'package-amd64: [rpm-x64]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'package-arm64: [deb-arm64]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
'package-arm64: [rpm-arm64]' job needs 'build-gem' job, but 'build-gem' is not in any previous stage
```

@alexjf figured out correctly that because we excluded the `build-gem`
step entirely, this made GitLab CI really unhappy.

We discussed a bit if we should still build the gem or not in
the nightly build, but ultimately decided to still let it build
(at least as a simplification to move forward on this issue).

**Motivation:**

Without this our nighly deploys are failing.

**Additional Notes:**

N/A

**How to test the change?**

We'll test a schedule run from this branch to see if it works fine.