[PROF-9370] Add microbenchmark coverage for GC profiling
**What does this PR do?**

As part of validating the overhead of the GC profiling feature, I've
created 4 sets of benchmarks to be able to characterize its
performance:

* `profiler_gc`: Measures the cost, on the profiler side, of each GC
  iteration
* `profiler_gc_minute`: Measures the cost of sampling one minute of an
  app with a lot of GC activity
* `profiler_gc_integration`: Measures the overhead of GC profiling on
  major GC cycles
* `profiler_gc_integration_allocations`: Measures the overhead of GC
  profiling on minor GC cycles

I'm preparing a writeup with analysis of these results, which I'll
share soon.

**Motivation:**

Validate overhead of GC profiling feature before we enable it by
default.

**Additional Notes:**

N/A

**How to test the change?**

Run the benchmarks!