[NO-TICKET] Add type signatures for `Profiling::Collectors::CodeProvenance`
**What does this PR do?**

This PR gets us slightly closer to the goal of having the profiler fully
type checked using `Steep`.

I've added type signatures for `CodeProvenance` and did a few tweaks
to the class to make it pass the type checks, since `Steep` was not
very happy with the way we were using a `Struct`.

**Motivation:**

Increase type checking coverage for profiler.

**Additional Notes:**

N/A

**How to test the change?**

Validate that the type checker passes + CI still green.