Introduce `CpuAndWallTimeWorker#wait_until_running` helper
Often in our testing and benchmarking, the need to wait for the profiler
to run comes up (so that for instance we don't miss some allocations,
if the profiler only actually starts running half-way through the test).

Thus far, we did this by looping + sleeping around
`_native_is_running?`, but I've decided to add a nicer API that avoids
all this confusion.