[PROF-9370] Enable Garbage Collection profiling by default
**What does this PR do?**

This PR turns the Garbage Collection profiling feature on by default.

Specifically it:

1. Introduces a new `c.profiling.advanced.gc_enabled` setting, which
   defaults to `true` (can also be toggled using the
   `DD_PROFILING_GC_ENABLED` env var)

2. Deprecated the previous
   `c.profiling.advanced.force_enable_gc_profiling` setting. This
   setting only served to turn the feature on, not turn it off, so it
   no longer makes sense to exist.

3. Does not turn on GC profiling on Rubies affected by
   https://bugs.ruby-lang.org/issues/18464 . Specifically, GC profiling
   can now only be turned on in Ruby versions 2.x, 3.1.4+, 3.2.3+ and
   3.3+ . This avoids impacting known-buggy Rubies.

**Motivation:**

We've done through validation that this feature is working nicely and
with low overhead, and it works great with the timeline view.

**Additional Notes:**

Regarding the impact of Ruby issue 18464, I explicitly chose to no
longer allow GC profiling to be enabled on affected Rubies.

I did this so as to simplify the implementation of the settings.

Because we wanted to enable this feature by default, having a way
to still enable it on these Rubies would mean either having another
option (e.g. keep `force_enable_gc_profiling` around?), somehow
introducing a tri-state value for the setting (like we do for the
`no_signals_workaround_enabled`), or fiddling with precedences to
distinguish the default value from a user-set one.

In the end, I decided to simplify and not do any of the above. We
can still change our minds later and introduce it based on user
feedback and needs, but unless a user really needs to be on
3.0.x for which there is no bug fix release, we really really
really really instead recomend moving to the latest patch release
of the 3.1.x or 3.2.x series.

**How to test the change?**

The change includes test coverage. You can also profile a simple
script and observe that GC profiling is included by default, without
any extra option having to be set.