[PROF-9370] Remove separate configuration for benchmarking GC profiling
**What does this PR do?**

This PR removes the benchmarking configuration for testing GC
profiling.

This is because in https://github.com/DataDog/dd-trace-rb/pull/3558
GC profiling was enabled by default and so the `only-profiling`
configuration now includes GC profiling (and thus we don't need
a separate one).

**Motivation:**

Remove benchmarking configurations that no longer make sense to have.

**Additional Notes:**

N/A

**How to test the change?**

Validate this configuration is no longer tested.