[NO-TICKET] Upgrade to libdatadog 8
**What does this PR do?**

This PR upgrades dd-trace-rb to use libdatadog 8.

Even though there were a few breaking API changes, none of those APIs
were being used in the Ruby profiler, so it's a very minor bump for us.

**Motivation:**

This version includes a number of fixes and tweaks that will finally
enable us to land the crash tracker feature
(https://github.com/DataDog/dd-trace-rb/pull/3384).

**Additional Notes:**

I'm opening this PR as a draft as libdatadog 8 is not yet available on
rubygems.org, and I'll come back to re-trigger CI and mark this as
non-draft once it is.

**How to test the change?**

Our existing test coverage includes libdatadog testing, so a green CI
is good here :)