[NO-TICKET] Remove profiler support for Ruby 2.3 and 2.4
**What does this PR do?**

This PR removes the code for supporting Ruby 2.3 and 2.4 from the
profiler.

This matches the minimum version for the datadog gem which is now
Ruby 2.5 (e.g. this was dead code -- you couldn't actually use it
anymore).

**Motivation:**

We opted not to do this earlier on in the 2.x release cycle,
but as we are getting close to the final release, and given that the
diff is not that big, I decided to go ahead and do it now.

**Additional Notes:**

In some cases, because a conditional was removed, indentation
was adjusted. I suggest reviewing this diff with "Hide whitespace"
turned on (or `-w` in the git cli).

In ddtrace 1.x, the gem supported Ruby 2.1+ whereas the profiler
was supporting 2.3+. We still had some awkward support code to
deal with this difference, which I also went ahead and removed.

It's cool in particular to see `private_vm_api_access.c` shrinking.
I'm looking forward to this trend continuing, and maybe one day we'll
even be able to remove it completely :)

**How to test the change?**

Validate that CI is still green on supported Rubies.