Add hanami app to integration CI configurations (#3639)
What does this PR do?

Adds hanami to CI configurations, as requested in #2546.

Several things needed to be fixed for the tests to pass:

Fix hanami integration app dependencies #2546
There was an empty hanami.rb which looks to have been intended to mock Hanami framework (?). The app requires the actual hanami framework to function and this empty file was loaded instead of hanami itself. I removed the file.
Hanami integration app was missing unicorn configuration, which the test suite was expecting. I copied it from sinatra.
The controller in the test app was written to fail 50% of the time, presumably to provide an environment to test failure. I commented out the failure generation so that CI would pass every time.
I also made the image build script work when ruby version isn't given and to provide a help message so that someone new to this area of the code could actually get it running locally.

Motivation:

Review of #2546.

Test plan
The tests are running on Ruby 2.6 and 2.7 only. Hanami 1 (1.3.5, more exactly) that is currently being used in these tests is not compatible with Ruby 3.


Co-authored-by: Oleg Pudeyev <code@olegp.name>
Co-authored-by: Tony Hsu <tonyc.t.hsu@gmail.com>