[NO-TICKET] Minor: Use fiddle to simulate crash in spec
**What does this PR do?**

This PR adds an additional test case to the profiling crashtracker
as discussed in
https://github.com/DataDog/dd-trace-rb/pull/3384#discussion_r1595866807
(thanks @lloeki for the suggestion).

**Motivation:**

Improve test coverage for the feature.

**Additional Notes:**

The diff looks more noiser than it is because of whitespace changes,
I recommend reviewing this PR without them.

**How to test the change?**

Check that the new test passes!