Fix matching of decoded span ids from profiles
Pprof decodes numeric labels as signed 64-bit values BUT as a slight
cheat both the profiler and the backend interpret them as unsigned
64-bit values so we can represent the full 64-bit values used for span
ids in traces.

But we were not accounting for this in our tests, so the decoded pprofs
we use for testing were not matching the expected values.

I've tweaked the pprof test reading code to account for this.