[NO-TICKET] Switch libdatadog-crashtracking-receiver checks to wait_for
**What does this PR do?**

This PR changes the crashtracking specs to use `wait_for` when asserting
on the `libdatadog-crashtracking-receiver` process still being alive or
not.

**Motivation:**

While experimenting with moving our test suite to GitHub Actions, this
seemed to trigger a lot.

Also, we saw it trigger at least once for CircleCI as well:
https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/15402/workflows/d3d6f9ad-7d9f-4845-9f6b-bc92eaf88b74/jobs/563014

**Additional Notes:**

Fixes https://github.com/DataDog/ruby-guild/issues/176 .

I'll see if there's something we can improve on the libdatadog side
to avoid this flakiness, but let's not live with the flaky specs until
we do so.

**How to test the change?**

Validate that CI is still green.