[PROF-10099] Upgrade to libdatadog 10
**What does this PR do?**

This PR upgrades the datadog gem to use libdatadog 10.

There was a tiny breaking API change, but otherwise it's a drop-in
upgrade.

**Motivation:**

There were a few improvements to crash tracking that I'd like to
pick up (some of the features need to be configured -- I'll open a
separate PR for it).

Also, staying on the latest libdatadog enables us to pick up the
latest improvements and also makes it easier to develop new features
that need libdatadog changes.

**Additional Notes:**

As usual, I'm opening this PR as a draft as libdatadog 10 is not yet
available on rubygems.org, and I'll come back to re-trigger CI and mark
this as non-draft once it is.

**How to test the change?**

Our existing test coverage includes libdatadog testing, so a green CI
is good here :)